+++ src/lib.rs
@@ -12,11 +12,13 @@ mod vec2_range_iter;
 mod vec3;
 mod vec3_range_iter;
 mod zero;
+mod normal_limits;
+mod large_type;
 
 pub use self::{
     cast::*, containing_float::*, float::*, goniometric::*, into_float::*,
     isqrt::*, one::*, scale::*, sqrt::*, vec2::*, vec2_range_iter::*, vec3::*,
-    vec3_range_iter::*, zero::*,
+    vec3_range_iter::*, zero::*, normal_limits::*, large_type::*,
 };
 
 #[cfg(test)]